Homework 18
Please give feedback because when I did exactly as instructed I failed the test and I had to do a long search on the internet and add a JavaScript executor to pass the test. We didn't study this and it's already just a search for options. Please tell me what my mistake is